Move ObjectVariableNotSetInspectionTests for QuickFixes into relevant files under RD.QuickFixes
