+++ RubberduckTests/Inspections/ObjectVariableNotSetInspectionTests.cs
@@ -2,7 +2,6 @@ using System.Linq;
 using System.Threading;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Rubberduck.Inspections.Concrete;
-using Rubberduck.Inspections.QuickFixes;
 using Rubberduck.Parsing.VBA;
 using Rubberduck.VBEditor.SafeComWrappers;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
@@ -291,120 +290,6 @@ End Function";
 
             Assert.AreEqual(expectedResultCount, inspectionResults.Count());
 
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void ObjectVariableNotSet_IgnoreQuickFixWorks()
-        {
-            var inputCode =
-            @"
-Private Sub Workbook_Open()
-    
-    Dim target As Range
-    target = Range(""A1"")
-    
-    target.Value = ""forgot something?""
-
-End Sub";
-            var expectedCode =
-            @"
-Private Sub Workbook_Open()
-    
-    Dim target As Range
-'@Ignore ObjectVariableNotSet
-    target = Range(""A1"")
-    
-    target.Value = ""forgot something?""
-
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new ObjectVariableNotSetInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-            
-            new IgnoreOnceQuickFix(state, new[] {inspection}).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void ObjectVariableNotSet_ForFunctionAssignment_ReturnsResult()
-        {
-            var expectedResultCount = 2;
-            var input =
-@"
-Private Function CombineRanges(ByVal source As Range, ByVal toCombine As Range) As Range
-    If source Is Nothing Then
-        CombineRanges = toCombine 'no inspection result (but there should be one!)
-    Else
-        CombineRanges = Union(source, toCombine) 'no inspection result (but there should be one!)
-    End If
-End Function";
-            var expectedCode =
-            @"
-Private Function CombineRanges(ByVal source As Range, ByVal toCombine As Range) As Range
-    If source Is Nothing Then
-        Set CombineRanges = toCombine 'no inspection result (but there should be one!)
-    Else
-        Set CombineRanges = Union(source, toCombine) 'no inspection result (but there should be one!)
-    End If
-End Function";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(input, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new ObjectVariableNotSetInspection(state);
-            var inspectionResults = inspection.GetInspectionResults().ToList();
-
-            Assert.AreEqual(expectedResultCount, inspectionResults.Count);
-            var fix = new UseSetKeywordForObjectAssignmentQuickFix(state);
-            foreach (var result in inspectionResults)
-            {
-                fix.Fix(result);
-            }
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void ObjectVariableNotSet_ForPropertyGetAssignment_ReturnsResults()
-        {
-            var expectedResultCount = 1;
-            var input = @"
-Private example As MyObject
-Public Property Get Example() As MyObject
-    Example = example
-End Property
-";
-            var expectedCode =
-            @"
-Private example As MyObject
-Public Property Get Example() As MyObject
-    Set Example = example
-End Property
-";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(input, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new ObjectVariableNotSetInspection(state);
-            var inspectionResults = inspection.GetInspectionResults().ToList();
-
-            Assert.AreEqual(expectedResultCount, inspectionResults.Count);
-            var fix = new UseSetKeywordForObjectAssignmentQuickFix(state);
-            foreach (var result in inspectionResults)
-            {
-                fix.Fix(result);
-            }
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
 
         [TestMethod]
