+++ RubberduckTests/RubberduckTests.csproj
@@ -133,6 +133,7 @@
     <Compile Include="QuickFixes\SpecifyExplicitByRefModifierQuickFixTests.cs" />
     <Compile Include="QuickFixes\SpecifyExplicitPublicModifierQuickFixTests.cs" />
     <Compile Include="QuickFixes\SplitMultipleDeclarationsQuickFixTests.cs" />
+    <Compile Include="QuickFixes\UseSetKeywordForObjectAssignmentQuickFixTests.cs" />
     <Compile Include="SmartIndenter\VerticalSpacingTests.cs" />
     <Compile Include="SourceControl\SynchrounouslyConstructedDeclarationFinderFactory.cs" />
     <Compile Include="Symbols\AccessibilityCheckTests.cs" />
